security/cryptopp: Add patch from pkgsrc
The definition of CRYPTOPP_BOOL_ALIGN16_ENABLED was incorrect for i386.
For NetBSD, this resulted in the inability to build cryptopp on i386.
On FreeBSD, it built, but deskutils/cdcat choked on it while building on
FreeBSD 10/i386 and FreeBSD 11/i386.

PR:		192199
Fix from:	pkgsrc security/crypto++
Discussed:	NetBSD GNATS 48134
Approved by:	delphi@


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@363448 35697150-7ecd-e111-bb59-0022644237b5