New port: devel/p5-MooX-HandlesVia
MooX::HandlesVia is an extension of Moo's 'handles' attribute
functionality. It provides a means of proxying functionality from
an external class to the given atttribute. This is most commonly
used as a way to emulate 'Native Trait' behavior that has become
commonplace in Moose code, for which there was no Moo alternative.

WWW: http://search.cpan.org/dist/MooX-HandlesVia/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@364218 35697150-7ecd-e111-bb59-0022644237b5