sysutils/hal: Adjust pkg-plist so it passes stage-qa and leftover checks
Hal was not passing poudriere testport or poudriere bulk -t runs due
to failing the stage-qa tests and later a leftover check with
/var/lib/PolicyKit/user-haldaemon.auths.  The reported error is actually
not wrong, but rather an inadequacy in the check-plist logic.  However,
hald is a dependency for a lot of ports and it keeps breaking bulk -t
runs, so I switched "@unexec rm -rf" to "@unexec rmdir" to that it would
build cleanly in poudriere (gnome@ had fair warning!).  If the check-plist
logic is improved in the future, this can be changed back to the better
"rm -rf" command then.

While here, stop trying to remove directories that PolicyKit installed
because this will always fail.  This clears up stage-qa warnings.

PR:		191223
Submitted by:	marino
Approved by:	maintainer timeout (2 months, 2 pings)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@364801 35697150-7ecd-e111-bb59-0022644237b5