security/razorback-officeCat: Chase json-c update (Unbreak)
This port is i386 only (due to dependency on wine, so the ONLY_FOR_ARCH
definition is actually unnecessary) so I can't test it.  It's also
unstaged.  However, the standard json-c patches apply, so just assume
that officeCat needs them.  If I'm wrong, the port will probably be
pruned in a week anyway.

Approved by:	portmgr (implicit, NO_STAGE)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@365504 35697150-7ecd-e111-bb59-0022644237b5