devel/rubygem-hashie: update to 3.2.0 to satisfy devel/rubygem-grape update
devel/rubygem-grape was updated to a newer version which requires a newer
version of hashie. The update doesn't seem to harm anything that depends on it,
so go ahead and update hashie. If we discover later we need the older hashie,
we can make a rubygem-hashie2 port.

With hat:	ruby@


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@366070 35697150-7ecd-e111-bb59-0022644237b5