Remove the --with-system-graphite configure option to unbreak editors/openoffice-4 on FreeBSD 10.0 and 11.0.  The cause of the breakage is that on systems with clang as the default compiler, graphics/silgraphite is built with clang and linked to libc++.so.1, while openoffice is built with gcc from ports and linked to libstdc++. This combination causes the resulting executables to go boom.
Add a source makefile patch to unbreak the --with-system-jpeg option
and enable this option.  The problem is that without this patch the
build tries to link to the openjdk version of libjpeg instead of
the version installed by graphics/jpeg.

The --with-system-mythes options also works, so enable it as well.

Adjust LIB_DEPENDS for these changes.  Also make the graphics/png
dependency explicit.

Reroll the ext_sources tarball for these changes.  Remove SVNREVISION
from its name and tweak how it is extracted to allow the same
ext_sources tarball to be used for multiple AOO releases.

Various Makefile tweaks to make upgrades easier and to reduce
differences between the Makefile for openoffice-4 and a forthcoming
update to openoffice-devel.

USE_PYTHON=yes is obsolete, so convert to USES=python.

Remove a bunch of @'s from the Makefile to make the build slightly
more verbose and make debugging a bit easier if something breaks
in our Makefile.

Remove trailing cruft from files/patch-sal (from pfg@).

Nuke files/patch-sal-inline (from pfg@).

PR:		192545
Approved by:	office@ (bapt)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@366163 35697150-7ecd-e111-bb59-0022644237b5