Remove CONFIGURE_ARGS=--enable-explicit-deps.  This option adds dependencies to the Requires field in pkgconfig files that should only be needed with static linking and cause a lot of overlinking with shared linking.  They should be added to the Requires.private field instead, but since this port doesn't install any static libraries they can just be removed.
The flag was added because several ports that depend on this port failed to
link because of missing libraries when the linker enforced explicit
linking.  The pkgconfig files in this port then provided the links.
These ports have all been fixed now.

PR:		192062
Exp-run by:	antoine
Approved by:	portmgr (antoine)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@366393 35697150-7ecd-e111-bb59-0022644237b5