Stage japanese/eijiro-fpw
I modernized and simplified this port beyond the PR, including embedding
the pkg-plist and converting patch-aa with DOS line endings and using
dos2unix.  I also removed the shell command at the small expense of
making lha a permanent dependency.  The downside is that this is a manual
package with a restricted cdrom, so I can't test any of it.  I'll commit
this with the hope that it's correct and request the submitter, who does
own the restricted data, verify that it's working as hoped.

PR:		193061
Submitted by:	turutani (kyoto)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@366566 35697150-7ecd-e111-bb59-0022644237b5