Update to 1.3.20.
While here, fix up the OPTIONS. X11 was used as an option and as a
category name, which made it very confusing what you could and
could not select. Fix this by simply telling the user they have to
enable X11 to get DPS.

Also, I have removed DPS from OPTIONS_DEFAULT. It brings in gtk12 and
frankly that is ridiculous in 2014.

I'm not touching the mysterious reassignment of CXX after <bsd.port.post.mk>
with a ten foot pole.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@366831 35697150-7ecd-e111-bb59-0022644237b5