Add security/p5-Crypt-Tea_JS.
security/p5-Crypt-T_e_a development stopped at 2.11 and there will be no
further releases for that port. Development has continued by the same
author, under a new name, Crypt::Tea_JS.

I've added conflicts between the two, but p5-Crypt-T_e_a should probably
be considered deprecated.

This module implements TEA, the Tiny Encryption Algorithm, and some Modes of
Use, in Perl and JavaScript.

The $key is a sufficiently longish string; at least 17 random 8-bit bytes for
single encryption.

Crypt::Tea_JS can be used for secret-key encryption in general, or, in
particular, to communicate securely between browser and web-host. In this case,
the simplest arrangement is for the user to enter the key into a JavaScript
variable, and for the host to retrieve that user's key from a database. Or, for
extra security, the first message (or even each message) between browser and
host could contain a random challenge-string, which each end would then turn
into a signature, and use that signature as the encryption-key for the session
(or the reply).

WWW: http://search.cpan.org/dist/Crypt-Tea_JS/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@366933 35697150-7ecd-e111-bb59-0022644237b5