Add new port net/cvsup-static
The ports lang/ezm3, net/cvsup and net/cvsup-without-gui ports have
expired and will be removed as soon as today.  This port provides
a statically-compiled version of cvsup-without-gui to offer cvsup on
the amd64 platform of all supported releases of FreeBSD.

This hasn't been available since ezm3 coredumps on newer releases, and
now it's BROKEN due to being unstaged.

PR:		192180
Submitted by:	delphij@


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@367724 35697150-7ecd-e111-bb59-0022644237b5