Add new port graphics/xcftools
PR:		179116
Submitted by:	Kalten (gmx.at)

Xcftools is a set of fast command-line tools for extracting information
from the Gimp's native file format XCF. The tools are designed to allow
efficient use of layered XCF files as sources in a build system that use
'make' and similar tools to manage automatic processing of the graphics.
These tools work independently of the Gimp engine and do not require the
Gimp to even be installed.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@368415 35697150-7ecd-e111-bb59-0022644237b5