astro/xtide: Add support for non-ports GCC
This port requires GCC, so it loads the lang/gcc.  The ports GCC will
look in /usr/local/include by default, but others (such as DF base
compiler) will not.  Specify -I${LOCALBASE}/include in CFLAGS explicitly
as it is needed to locate the X11/Xaw/Box.h header.  This is covered by
"Just Fix It" blanket.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@369179 35697150-7ecd-e111-bb59-0022644237b5