lang/ruby2*: fix dtrace support
DTrace support in ruby needs the -xnolibs flag added to the dtrace command
called during build to ensure that dtrace support is enabled even if the dtrace
kernel modules are not loaded at build time.

However, with this dtrace will be enabled on some OS versions where it will not
work, so add logic to disable it where appropriate.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@369732 35697150-7ecd-e111-bb59-0022644237b5