New port: security/py-rekall
The Rekall Framework is a completely open collection of tools, implemented in
Python under the GNU General Public License, for the extraction of digital
artifacts from volatile memory (RAM) samples. The extraction techniques are
performed completely independent of the system being investigated but offer
visibility into the runtime state of the system. The framework is intended to
introduce people to the techniques and complexities associated with extracting
digital artifacts from volatile memory samples and provide a platform for
further work into this exciting area of research.

WWW: http://www.rekall-forensic.com/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@370110 35697150-7ecd-e111-bb59-0022644237b5