devel/rubygem-mustache: create port
Inspired by ctemplate, Mustache is a framework-agnostic way to render
logic-free views.

As ctemplates says, "It emphasizes separating logic from presentation:
it is impossible to embed application logic in this template
language.

Think of Mustache as a replacement for your views. Instead of views
consisting of ERB or HAML with random helpers and arbitrary logic,
your views are broken into two parts: a Ruby class and an HTML
template.

WWW: http://github.com/defunkt/mustache


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@370140 35697150-7ecd-e111-bb59-0022644237b5