Add new port emulators/hyperv-is (FreeBSD 8.4, 9.1, 9.2, 9.3, 10.0)
PR:		182209
Submitted by:	kylie (bsdic @ microsoft)

The hyperv-is provision a collection of kernel mode drivers as well as
user-space daemons to facilitate integration with Hyper-v to provide a
feature rich and high performance FreeBSD guest experience.

The FreeBSD Integration Service on Hyper-v includes a collection of kernel
mode drivers as well as user-space daemons to interact with the drivers
that are required to run Hyper-V-specific devices known as FreeBSD
Integration Services (BIS). It is to facilitate integration with Hyper-v
to provide a feature rich and high performance FreeBSD guest experience.
See the man page for a list of binaries and their functions.

FreeBSD support for hyperv-is was first added by Microsoft BSD Integration
Services Team <bsdic@microsoft.com>.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@370242 35697150-7ecd-e111-bb59-0022644237b5