New port opengrm ngram.
The OpenGrm NGram library is used for making and modifying n-gram language
models encoded as weighted finite-state transducers (FSTs).  It makes use of
functionality in the OpenFst library to create, access and manipulate n-gram
models.  Operations for counting, smoothing, pruning, applying, and
evaluating models as well as support for distributed computation are among
those provided.  It was developed by contributors from OHSU and Google Research.

http://www.openfst.org/twiki/bin/view/GRM/NGramLibrary


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@371272 35697150-7ecd-e111-bb59-0022644237b5