The IRC configure script has curious logic, is being overly complex, and actually breaks badly if the port is being built in a chroot where the host triplet won't match the jail triplet.
The solution I've provided here is clean.  Using the pre-configure
target, the build directory (with a fixed name) is created and the
needed configure script and header are copied over, just like the top
configure script does.  Then the port makefile just configures with
that.  It's much better, more robust, and fixes the triplet mismatch.

PR:		194679
Submitted by:	marino


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@371671 35697150-7ecd-e111-bb59-0022644237b5