bsd.port.mk: Finish update to make makepatch
A portion of this patch to upgrade makepatch was committed almost 2
months ago; this is the rest of it.  It changes the directory separator
to "_" and it will transform "_" in the filename to "__" to avoid
ambiguous file names (e.g. A/B/C.c and A_B/C.c won't have the same patch
name).

The new logic will not rename an existing patch that used previously
standard separators of "-", "+", or "__" in its name.  It is desireable
to avoid commits that only change the filename of the patch, so that's
why existing filenames are re-used if previously legal.

The diff command is also pass the -p argument for additional useful
context.

Differential Revision:	https://reviews.freebsd.org/D582
Approved by:		portmgr (bapt)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@371776 35697150-7ecd-e111-bb59-0022644237b5