www/davical: Reorganize port to fix DOCS option, Fix Mac OS contacts.app
The DOCS option dumped the critical dba directory along with non-docs
directories scripts and config into DOCSDIR, which resulted in a broken
port if the user turned off the DOCS option.

Reorganize the port to put dba and scripts in DATADIR, and config in
EXAMPLESDIR and add the EXAMPLES option.  While here, use the @sample
keyword, clean up @dirrm, and tweak a bit.

The second PR fixes vcards on DAVICal servers for Mac OS 10.9.

PR:		187112
PR:		186502


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@371860 35697150-7ecd-e111-bb59-0022644237b5