Add new port databases/pecl-yac
PR:		194553
Submitted by:	John Pupu

Yac is a shared memory user data cache for PHP.  It can be used instead of
APC or local memcached.

Yac is lockless, which it is very fast, but there could be a chance the
wrong data will be retrieved (it depends on how many key slots are
allocated and how many keys are stored), so a product using Yac should not
be very sensitive to data loss.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@372163 35697150-7ecd-e111-bb59-0022644237b5