deskutils/griffith: Set LXML option by default
This port is unmaintained.  The referenced PR indicates that LXML
dependency is not actually optional and run-time files if it's not
installed.  If true, it should be installed unconditionally, but as
a half-measure, I'll just switch the option to fix it for standard
binary packages, and let the future maintainer look into this more.

PR:		194151
Submitted by:	ktullavik (gmail)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@372410 35697150-7ecd-e111-bb59-0022644237b5