databases/cego: 2.20.16 -> 2.20.17
Change for primary and secondary host evaluation.

Instead of comparing gethostbyname, the defined dbhost value is used.
This helps if the configured dbhost does not match with the result
value of gethostbyname (e.g. is multiple hostnames are configured
on a database server)

Submitted by:	Björn Lemke <lemke@lemke-it.com>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@372626 35697150-7ecd-e111-bb59-0022644237b5