The FreeBSD GNOME team proudly presents GNOME 3.14 and Cinnamon 2.2. Gnome 3.14.1 and Cinnamon 2.2.16 are supported on FreeBSD 9.3-RELEASE and up.
This commit removes the old GNOME 2 desktop, bindings and some ports that
can't be compiled. A few ports where updated to more recent versions to
allow them to compile with this update.

Apart from updating ports to newer versions

GDM is more integrated with gnome-shell now, and handles several things for
the GNOME desktop such as screen locking. If you want to use GNOME 3 via
startx, you will have to add your own lock screen/screensaver. For example xscreensaver
can be used for sessions started without GDM.

Shell Extensions can be installed via https://extensions.gnome.org/ , we have
ported a few that can't be installed via this way.
The old gnome-utils and gnome-games ports where split up into single ports
and where converted to meta-ports.
gnome-terminal requires a UTF-8 locale to run, gdm handles this already, but
if you use startx you need to do this yourself.

Upgrade instructions:

Delete the old and conflicting packages:
# pkg delete clutter gnome-utils gnome-panel gnome-keyring vala-vapigen \
    guile gcalctool gnome-media libgnomekbd
# pkg delete gnome-screensaver gnome-applets bug-buddy evolution-exchange \
    evolution-webcal gnome-system-tools seahorse-plugins gnome-control-center

For package users the following lines will be enough:
# pkg upgrade
# pkg install gnome3

For ports users should do the following:
# portmaster -a
# portmaster x11/gnome3

We are currently aware of two issues. The first issue is a bug in the
file monitoring code in the glib20 port. This bug causes glib programs
to crash when files in a monitored directory are added or removed.
Upstream is aware of the problem, but since the problem is quite complex
there is no solution yet. This problem isn't restricted to BSD.

The second issue is that on certain video cards totem will display a
purple/pink overlay on the video. It not clear yet where the issues
comes from.

Major thanks goes to Gustau Perez for being a driving force behind getting
GNOME 3 up to speed again. Also thanks to Antoine Brodin for running the exp-runs.

This update was also made possible by:
	Joe Maloney
	Kris Moore
	Beeblebrox
	Ryan Lortie
	Antoine Jacoutot
	and everyone I missed


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@372768 35697150-7ecd-e111-bb59-0022644237b5