Update to 3.1.0.
19 months have passed since the release of 3.0.0 and there have been 1259
commits from over 24 contributors. Though many of the bugfixes did make it into
the 3.0 branch, not all of them did.

3.1 brings many speedups, new features and commands, a new web interface, and
build hooks. The new web interface works at a top-level, jail-level and the
build-level still by using a static site with AJAX.

There is a queue system that is not quite ready yet but should make it into
3.2.

The release notes documenting the major changes is at
https://github.com/freebsd/poudriere/wiki/release_notes_31

Thank you to everyone involved.

Support can be found in #poudriere on Freenode.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@373856 35697150-7ecd-e111-bb59-0022644237b5