Add a port of libxo:
libxo - A Library for Generating Text, XML, JSON, and HTML Output

The libxo library allows an application to generate text, XML, JSON,
and HTML output using a common set of function calls. The application
decides at run time which output style should be produced. The
application calls a function "xo_emit" to product output that is
described in a format string. A "field descriptor" tells libxo what
the field is and what it means.

Sponsored by:	DARPA, AFRL


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@374819 35697150-7ecd-e111-bb59-0022644237b5