net-mgmt/zabbix2[24]-server: fix zabbix and database startup sequence
The Zabbix server and proxy ports require a database to run. They
currently support four, PostgreSQL, Oracle, SQLite, and MySQL. As
of now, rcorder puts them ahead of the database ports, so they spit
out dozens of error messages during startup and, worse, get the
database pulled out from under them during shutdown.

The ports used to provide a solution for this, by defining
ZABBIX_REQUIRE. The value of this variable used to be added to the
REQUIRE lines in the startup scripts, but this was removed accidentally
from 2.2 in r355546 and never existed in 2.4. This patch adds it
back, in a slightly different, optionsng-friendly way.

PR:		195081
Submitted by:	chris@chrullrich.net
Approved by:	pakhom706@gmail.com (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@375683 35697150-7ecd-e111-bb59-0022644237b5