New port: devel/p5-Object-Import
This module lets you call methods of a certain object more easily
by exporting them as functions to a package. The exported functions
are not called as methods and do not receive an object argument,
but instead the object is fixed at the time you import them with
this module.

WWW: http://search.cpan.org/dist/Object-Import/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@376150 35697150-7ecd-e111-bb59-0022644237b5