change command_interpreter from /usr/bin/perl to ${PREFIX}/bin/perl
Several ports had rc.d scripts with hardcoded command_interpreter string
as /usr/bin/perl.  This symlink is not guaranteed to be in place, and it
isn't even an option for perl 5.20.  For affected ports, the interpreter
was changed to localbase.

In one case, the interpreter was correct, but it wasn't surround by
quotes.  Since the rc.d script would break if a space was contained in
${PREFIX}, quotes were added in that case.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@376298 35697150-7ecd-e111-bb59-0022644237b5