Add devel/gitinspector.
gitinspector is a statistical analysis tool for git repositories. The
defaut analysis shows general statistics per author, which can be
complemented with a timeline analysis that shows the workload and
activity of each author. Under normal operation, it filters the
results to only show statistics about a number of given extensions and
by default only includes source files in the statistical analysis.

Differential Revision:	https://reviews.freebsd.org/D1529
Reviewed by:	bapt
Approved by:	bapt


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@377097 35697150-7ecd-e111-bb59-0022644237b5