Add sysutils/qt5-qtpaths.
This is a command-line interface to QStandardPaths, part of qttools. It is
used by the KDE Frameworks 5 ports, which will be added in the future.

While here, also add x11-toolkits/qt5-quickcontrols to devel/qt5 which I had
forgotten to do before.

Submitted by:	Tobias Berner <tcberner@gmail.com>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@378033 35697150-7ecd-e111-bb59-0022644237b5