[NEW] www/py-autobahn: WebSocket library, WAMP real-time framework
Autobahn|Python is a networking library that is part of the Autobahn
project and provides implementations of:

  * The WebSocket Protocol
  * The Web Application Messaging Protocol (WAMP)

for Twisted and asyncio, on Python 2 & 3 and for writing servers and
clients.

WebSocket allows bidirectional real-time messaging on the Web and WAMP
adds asynchronous Remote Procedure Calls and Publish & Subscribe on top
of WebSocket.

WWW: http://autobahn.ws/python


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@378952 35697150-7ecd-e111-bb59-0022644237b5