devel/libssh: Remove -pedantic-error flags and reset port
This port breaks if it violates standards according to the -pedantic flag.
Accord to gcc-5, it indeed does with dozens of warnings like this:

/wrkdirs/security/libssh/work/libssh-0.6.3/include/libssh/priv.h:200:33:
  error: ISO C does not support '__FUNCTION__' predefined identifier
  [-Wpedantic] _ssh_set_error(error, code, __FUNCTION__, __VA_ARGS__)

The maintainer has been MIA for over a year with PRs on 8 ports.  I'm not
waiting for this one to time out too.  While here, reset MAINTAINER.

PR:		197785
Submitted by:	marino
Approved by:	maintainer reset


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@379559 35697150-7ecd-e111-bb59-0022644237b5