security/libgpg-error: Revert use of tradcpp (F11 regression)
With the previous change, libgpg-error does not pass configure target
on FreeBSD 11 due tradcpp failing "sanity checks".  This is a regression
since this port builds fine on FreeBSD 10.  However, 2037 ports get
skipped (this port is *way* too important) as a result, so revert the
gcc-5 fix until a solution can be found.

PR:		197562


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@380217 35697150-7ecd-e111-bb59-0022644237b5