Add new port lang/gcc5-aux (next Ada compiler)
This is the initial version of gcc5-aux, which will eventually become the
default Ada compiler.  It's not hooked into Mk/Uses/ada.mk yet, but it
does pass all Ada tests on both DragonFly64 and FreeBSD64.

For FreeBSD 10 amd64, the compiler built fine outside of poudriere but the
bootstrap compiler failed inside of it (seemingly as a result of using
base linker).  Eventually a new bootstrap compiler needs to be made, but
for now gcc5-aux is built with a full bootstrap on FreeBSD.  On DragonFly,
only a single stage is built as the bootstrap compiler still works fine.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@381294 35697150-7ecd-e111-bb59-0022644237b5