security/libsparkcrypto: Fix gcc5-aux support and stop stage building
This port failed to build on gcc5-aux because 3 data types could not
be inlined after an explicit pragma requested it (the pragmas were
removed).

I noticed half the port was building in the stage phase, so I changed
the ALL_TARGET to make all of it build during the build phase.  This
probably didn't need revbump, but I'm going to anyway to make sure the
FreeBSD builders are still happy.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@381430 35697150-7ecd-e111-bb59-0022644237b5