finance/libobx: Fix numerous issues
I came to fix only one problem with libofx, but when it failed to pass
stage-QA checks, I found others.

  1) CXXFLAGS not propagated (PHB 12.9) due to bug in configure (blanket)
  2) DragonFly iconv support (blanket, taken from DPorts)
  3) DOXYGEN option was no implemented
     * --disable-doxygen was set rather than DOXYGEN_CONFIGURE_ENABLE
     * DOXYGEN_BUILD_DEPENDS was not set
  4) disable doxygen setting, html documentation got installed
     unconditionally, but this was not reflected in plist

I decided that --disable-doxygen was the correct option because html files
were already provided an installed, so the DOXYGEN option was removed.  I
used PORTDOCS to fix the pkg-plist.  This port passes stage-QA now.

Approved by:	blanket and just-fix-it


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@381727 35697150-7ecd-e111-bb59-0022644237b5