games/airstrike: put this thing out of it's misery
This "game" was never completed, it was never playable, and now it looks
like it's not even being developed "slowly".  The respository and dev wiki
don't resolve (dyndns.org).  In my opinion, it never had the quality to
enter ports.  Since it's explicitly unmaintained, I'm setting it to be
removed on 10 April 2015.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@382296 35697150-7ecd-e111-bb59-0022644237b5