net/netembryo: Add 2 new options, SCTP and IPv6
DragonFly does not support SCTP.  This was detected fine in the configure
script, but check-plist failed due to an SCTP file that didn't get
installed.  The configure script takes options for SCTP and IPv6, so turn
those into options.  Use the brand-new OPTIONS_EXCLUDE_DragonFly feature
to allow SCTP to be on by default for FreeBSD and not even an option for
DragonFly.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@382335 35697150-7ecd-e111-bb59-0022644237b5