lang/gcc(46,47,48,49,5): Use OPTIONS_EXCLUDE_DragonFly to block JAVA
The JAVA frontend doesn't build on DragonFly on any release.  The new
OPTIONS_EXCLUDE_${OPSYS} feature is a nice way to avoid the use of
Makefile.DragonFly (most are in dports, but one is in lang/gcc5).

The recent addition of CXXFLAGS to lang/gcc5 prevents Makefile.DragonFly
on lang/gcc5 from being removed outright.  There are a couple of options
available to allow its removal, but I'll need to discuss with Gerald.

Approved by:	DragonFly blanket


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@382341 35697150-7ecd-e111-bb59-0022644237b5