Introduce FreeBSD xen dom0 components.
Add xen kernel and xen toolstack ports to the ports tree
This packages are for now only supported on FreeBSD 11 after 1100055
The xen package only include the kernel and an addition to the loader menu to
allow to disable booting on the xen kernel

The Xen toolstack (xl) includes seabios (hence the addition of LGPLv3 license)

Explicitly use nm from ports binutils to workaround a bug in FreeBSD's nm on
some version of FreeBSD.

Workaround configure checks that are actually not needed during the build of the
tool stack.

Thanks AllanJude@ for the xen.4th
Thanks roger@ for all the work making FreeBSD dom0 support working.

Please notice that you will need an Intel box with EPT and a working IOMMU


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@382965 35697150-7ecd-e111-bb59-0022644237b5