biology/platon: Upgrade from 05 MAR 2015 => 30 MAR 2015 to unbreak
As noted by previous commit, upstream rerolls the same distfile
frequently.  When FreeBSD cache was removed, the port broke.  I'm fixing
it again with a new digest, but I'm also defining DIST_SUBDIR, which is
the correct thing to do with rerolls-by-policy.

Also I removed the bad backup MASTER_SITES and tested in poudriere.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@383810 35697150-7ecd-e111-bb59-0022644237b5