p5-Wx: Unbreak on some platforms
This port was using CFLAGS for c++, so the -std=c++11 flag wasn't getting
passed even when USE_CXXSTD was set.  This port broke in some places when
c++11 became a requirement of Wx.

This is still broken on FreeBSD 8 (F9 not tested) due to new problems
related to compiling with gcc48.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@384341 35697150-7ecd-e111-bb59-0022644237b5