Add port for QtAwesome -- a small library, which makes it easier for QtApplications to use the "fontawesome" collection of icons.
Although the upstream source bundles the "font" itself with its sources
too, the port makes no use of it -- there are already about 2 scores
of fontawesome-FOO.ttf listed in various pkg-plist files in the tree.

This port's pkg-message will advise the user, that for the library to
be useful, the font-file must be known to fontconfig.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@384386 35697150-7ecd-e111-bb59-0022644237b5