Convert remaining p5-CGI.pm consumers to p5-CGI, and bump PORTREVISION.
All these changes are tested but the following are worth noting:

The following ports fail "make test", but did so before this change, and
fail in the same places:
- textproc/p5-xmltv
- www/p5-Business-Paypal
- www/p5-CGI-Enurl

www/p5-Apache-Gallery is missing all sorts of dependencies, fails all tests,
and should probably be marked BROKEN

For graphics/imc, move the OPTIONS_DEFINE block out of the LICENSE block area.

As long as we're here, sort plist on perl@-owned ports.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@384736 35697150-7ecd-e111-bb59-0022644237b5