New port: www/itop
iTop stands for IT Operational Portal.
iTop is an Open Source web application for the day to day operations
of an IT environment. iTop was designed with the ITIL best practices
in mind but does not dictate any specific process, the application
is flexible enough to adapt to your processes whether you want rather
informal and pragmatic processes or a strict ITIL aligned behavior.

WWW: http://www.combodo.com/

PR:		ports/199646
Submitted by:	olevole@olevole.ru


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@384886 35697150-7ecd-e111-bb59-0022644237b5