lang/gcc5: Remove Dragonfly's Makefile; it's no longer needed
Now that the JAVA exclusion is part of the main Makefile, the only thing
Makefile.DragonFly was doing was supporting pre-gcc5 dev branch.  Those
users have a static package available to them, and in general they are
expected to upgrade to post-gcc5 in base anyway.  Thus the DF makefile
is no longer necessary.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@384888 35697150-7ecd-e111-bb59-0022644237b5