databases/rocksdb: Unbreak on DragonFly due to -Werror (again)
The -Werror flag reappeared on the latest version, and it still doesn't
build without error on GCC5.  In this case, the FLAGS_env variable was
defined and not used tripping [-Werror=unused-variable]

I would seriously remove -Werror from this port (again), but this patch
makes it build again on DragonFly, at least for the moment.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@385116 35697150-7ecd-e111-bb59-0022644237b5