Revert r386285, removing exists() check on RUN_DEPENDS, for now.
The ports framework properly handles this but the tools that are used do not.
Portmaster and Portupgrade both incorrectly install openssh-portable even if
ssh-agent is installed already. Poudriere does not install or add a dependency
on it, but it does build the dependency.

At least the portmaster and portupgrade issues must be resolved first. They
will also fix any other port installing unneeded dependencies and allow
many of the exists() checks to come out of the tree.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@386347 35697150-7ecd-e111-bb59-0022644237b5