sysutils/rdate: Unbreak DragonFly after previous commit
OSVERSION was used without OPSYS.  This fix really should be
"if dragonfly or (freebsd osversion < X)" but currently OSVERSION
is valid as long as OPSYS <> DragonFly so this change will work and
is simpler than the "real" fix.  Ideally all of this should be replaced
with a feature detection mechanism.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@386393 35697150-7ecd-e111-bb59-0022644237b5