Add security/keepass, the official KeePass client.
KeePass is a free open source password manager, which helps you to manage your
passwords in a secure way. You can put all your passwords in one database, which
is locked with one master key or a key file. So you only have to remember one
single master password or select the key file to unlock the whole database. The
databases are encrypted using the best and most secure encryption algorithms
currently known (AES and Twofish).

This is the official KeePass application, which was originally only available
for Windows, but has now been re-written with .Net and able to be run on BSD,
Linux and Mac OS X with Mono.

WWW: http://keepass.info/

PR:		200062
Submitted by:	Ben Woods


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@386557 35697150-7ecd-e111-bb59-0022644237b5