Add new port: security/py-crits
CRITs is a web-based tool which combines an analytic engine with a cyber threat
database that not only serves as a repository for attack data and malware, but
also provides analysts with a powerful platform for conducting malware
analyses, correlating malware, and for targeting data. These analyses and
correlations can also be saved and exploited within CRITs. CRITs employs a
simple but very useful hierarchy to structure cyber threat information. This
structure gives analysts the power to 'pivot' on metadata to discover
previously unknown related content.

Approved by: mat (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@387056 35697150-7ecd-e111-bb59-0022644237b5