net/polyorb: convert to pkg-plist, support testsuite, plus
The pkg-plist isn't absurdly long, especially when docs are omitted, so
convert port from using generated plist to a static one with PORTDOCS
defined.  While here, allow the possibility of not installing docs
although this saves no effort and is not recommended since it removes
polyorb from GPS (Gnat Programming Studio IDE) menus.

Also import Makefile.in patch to properly strip binaries.

The main reason for the update was to support the ability to run the
testsuite.  When selected (off by default), the "run-testsuite" target
is available, and is intended to be executed after "make install".  The
testsuite target will build the examples along with the tests to be run,
but the examples are not installed.  However, additional .ali files are
installed as a result (before only their source files were installed).


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@387710 35697150-7ecd-e111-bb59-0022644237b5