cad/ghdl: Add option to use experimental LLVM backend
The LLVM backend builds much faster assuming clang35 and llvm35
are already available.  It probably builds on FreeBSD's base clang but
not ports llvm36.  In any case, the alternative backend does not
pass the testsuite on FreeBSD (fails at test 825).

The LLVM back is the default backend for DragonFly which allows the
port to build, but it fails the very first test (fails to lock mutex).
There's no need to bump -- for FreeBSD this should package the same
as it did before.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@388183 35697150-7ecd-e111-bb59-0022644237b5