The rencode module is similar to bencode from the BitTorrent project. For complex, heterogeneous data structures with many small elements, r-encodings take up significantly less space than b-encodings.
This version of rencode is a complete rewrite in Cython to attempt to
increase the performance over the pure Python module written by Petru
Paler, Connelly Barnes et al.  Later, it was forked, enhanced, and
bundled with Deluge.  Now, it is re-packaged and distributed by Xpra.

WWW: https://github.com/aresch/rencode


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@389078 35697150-7ecd-e111-bb59-0022644237b5