Add new port lang/lua-ada
These are bindings to a Lua library for Ada.  The gpr file is written for
the version of Lua that is default on the system that builds the port.
(see Mk/bsd.default-versions.mk)

There is no documentation available, but two examples with sources are
built and installed for illustration.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@389905 35697150-7ecd-e111-bb59-0022644237b5