Add new port: graphics/copperspice
I would consider copperspice "experimental" for the time being.  The
developers tested it on linux, windows, and macos, but not BSD.  Several
issues have been identified already:

 * differences in BSD install from GNU install caused fatal build
   failure (already fixed in v1.0.2)
 * It's not jobs safe (not reproducible on Linux, I suspect this is
   another difference of the install programs
 * specs don't exist for any BSD
 * qmutex was implemented for non-linux, non-windows
 * libtool is missing --tag
 * Makefile template missing dozens of lines to support OSTYPE_FREEBSD
   (or OSTYPE_BSD ?) which also needs to be implemented
 * FreeBSD 9 (and presumably 8) fails with "to_string" is not recognized
   as std namespace despite using gcc48 to build
 * FreeBSD10/i386 fails because of c++11 narrowing rules violations

The upsteam is already aware of some of these issues and is committed to
make CopperSpice work on BSD as a first class citizen.  So this initial
import might be considered a work in progress. I've been working on a
jobs-safe problem workaround but hit another jobs-safe failure.

=====

Copperspice is a C++ library derived from the existing Qt 4.8 framework.
The goal was to change the core design of the libraries, leveraging
template functionality and C++11 capabilities.

The redesign allowed the Qt Meta-Object Compiler (moc) system to be
completely removed.  Moc is a code generator and does not support many
aspects of C++ including templates, complex data types, static type
checking, and relies heavily on string comparisons. Removing moc improves
runtime performance, reduces the complexity of the build process, and
allows more issues to be detected at compile time.

Key features:
  * Qt Meta-Object Compiler (moc) is obsolete
  * Written in C++11
  * Library links directory to any standard C++ application
  * A template class can inherit from QObject
  * Copperspice includes several Qt5 classes


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@390166 35697150-7ecd-e111-bb59-0022644237b5