New port: devel/cloudabi-binutils.
This port installs a copy of GNU Binutils that targets CloudABI.
CloudABI is a compact UNIX-like runtime environment built on
capability-based security. More details:

	https://github.com/NuxiNL/cloudlibc

Support for CloudABI has been upstreamed, but will only be part of
the upcoming version of Binutils. This port includes a backported copy
of the patchset. It can safely be removed if devel/binutils is bumped to
a newer version.

PR:		200968
Reviewed by:	bapt


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@390208 35697150-7ecd-e111-bb59-0022644237b5