Add cloudabi-clang.
This port adds the necessary symbolic links to make Clang work as a
cross compiler for CloudABI. Clang can automatically derive its target
from argv[0]. We don't need to install another copy of Clang. It is
sufficient to invoke it the right way.

CloudABI support is not yet part of a released version of Clang. This is
why we still need to depend on llvm-devel. My intent is to stick to a
fixed compiler version at one point in time.

PR:		200954
Differential Revision:	https://reviews.freebsd.org/D2885
Approved by:	bapt


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@390358 35697150-7ecd-e111-bb59-0022644237b5