In April/May 2012, STEP Class Library was renamed to STEPcode. This was done because the old name wasn't accurate - the class libraries are only a fraction of the software.
The STEP Class Library (SCL) originated at the National Institute of Standards
and Technology, or NIST. NIST started working with STEP in the 80's and
continued until the late 90's. Some components of SCL were originally written
in Lisp and then re-written in mixed C and C++ in the early 90's.
The rest of SCL was written in C++ to begin with.

STEPcode (SC) includes the class libraries, some of the most widely used EXPRESS
schemas, some tools to work with EXPRESS, and support libraries for those tools.
Two of the tools can create schema-specific libraries that are used with the
class libraries. There are also some test files and programs.

WWW: https://github.com/stepcode/stepcode/wiki

PR:		201046
Submitted by:	fernando.apesteguia@gmail.com


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@391246 35697150-7ecd-e111-bb59-0022644237b5