www/squid: Support DragonFly SHM segments
Out of the box, squid would not run on dragonfly due to its handling
of SHM segments.  On DragonFly, SHM segments are always treated as files
but on FreeBSD it depends on whether or not application is inside a jail.

In any case, the case for DragonFly was no supported, so it has been
added via patch.  This also requires the return of /var/run/squid
directory which is where the SHM files are stored (defined by
localstatedir and supported by RC script).  The RC script would define
this directory if missing, but let's make sure it is always available.

PR:		201405
Submitted by:	marino
Approved by:	maintainer (timp87/gmail)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@391555 35697150-7ecd-e111-bb59-0022644237b5