graphics/copperspice: Mark broken everywhere
On FreeBSD clusters, this won't build (but it builds on my machine)
On DragonFly, the two removed TypeTrait.h patches are still required
(apparently this is a gcc thing, clang was happy enough without them)
Upstream is aware and working on it.

I may change this to BROKEN_FreeBSD when the missing patches are returned.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@391980 35697150-7ecd-e111-bb59-0022644237b5