- New port: www/bareos-webui
The bareos-webui is a free software tool written in PHP, intended to monitor
and manage Bareos over the web.

The webui is based on Zend Framework 2, an open source framework for developing
web applications and services using PHP 5.3+ and makes extensive use of the
Twitter Bootstrap front-end framework.

The bareos-webui is licensed under AGPL Version 3.

WWW: http://www.bareos.org/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@393760 35697150-7ecd-e111-bb59-0022644237b5