Introduce an option to build the new Qt5 based GUI. The Qt5 version will replace the now deprecated WXGTK version in the future. If you're currently launching 'mmg' (the WXGTK GUI) you will get the following notification:
  Please note that this program (mkvmerge GUI) is deprecated and will be removed
  within one or two releases. It has been superseded by the new MKVToolNix GUI
  which should also be included in the same package.

Don't enable this option by default to give current users a notice that future
updates will pull in the Qt5 dependency chain by default. However, users are
encouraged to try the Qt5 version and report bugs.

This new option does not change the current default build of the port and
therefore no PORTREVISION bump is required.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@394068 35697150-7ecd-e111-bb59-0022644237b5