percona56-client: Switch to USES=libedit.
Instead of manually adding both a build- and a run-time dependency on
libedit from ports, use USES=libedit instead. Not only does it take care of
adding the proper dependency, but it also sets CMAKE_PREFIX_PATH to
${LOCALBASE} to make sure that the version in ports is preferred instead of
the one in base.

This is a dependency for moving to CMake 3.3.x, as starting with 3.3.0 calls
to find_library() will also look in the $PATH environment variable and end
up finding libedit from base by default.

Approved by:	flo (maintainer)
Differential Revision:	https://reviews.freebsd.org/D3468


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@395190 35697150-7ecd-e111-bb59-0022644237b5