ports-mgmt/genpatch: Fix evaluation of WRKDIRPREFIX
The logic to define the PORTWORK variable evaluated WRKDIRPREFIX in some
cases, but this variable was never defined.  We need to pull in
bsd.port.pre.mk or bsd.port.options.mk to define it, and I chose the
latter.  This fixes the wrong value of PORTWORK being used when genpatch
is built manually (it was correct when built from poudriere).


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@395397 35697150-7ecd-e111-bb59-0022644237b5