Add upstream patch to fix the build with GCC 4.6, 4.7 and 4.8.
runetype.h uses _Thread_local, and if we pass -std=c11 or -std=gnu11 to GCC
the header expects it to be supported as part of the language and does not
make it a typedef or something else.

Since GCC only started supporting _Thread_local with the 4.9 series,
building CMake with, say, lang/gcc (which is 4.8) fails:

    /usr/include/runetype.h:92:22: error: expected '=', ',', ';', 'asm' or '__attribute__' before 'const'
    /usr/include/runetype.h: In function '__getCurrentRuneLocale':
    /usr/include/runetype.h:96:6: error: '_ThreadRuneLocale' undeclared (first use in this function)
    /usr/include/runetype.h:96:6: note: each undeclared identifier is reported only once for each function it appears in

The upstream patch adds a test for _Thread_local and uses C99 instead of C11
if it fails.

PR:		203066


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@397068 35697150-7ecd-e111-bb59-0022644237b5