Update Mesa port to 10.6.6 and add Clover. Add beignet 1.1.0. Add clinfo, clblas, clfft and clrng.
The major change is that all Mesa ports are now configured the same way.
This fixes several problems and enables new features.  The details
are described in this blog post:
http://blogs.freebsdish.org/graphics/2015/03/18/unifying-mesa-ports-configure/

The second important change is the OpenCL support.  Mesa's
implementation, Clover, is enabled as well as Beignet.  Clover
targets all Gallium drivers, only Radeon GPUs in our case.  Beignet
is for Intel GPUs starting with Ivy Bridge.  Thanks to Johannes
Dieterich, O. Hartman, and Koop Mast for their work on OpenCL!  As a
bonus, there are several OpenCL-based math ports added (clblas,
clfft and clrng). For more information and known issues, please see
https://wiki.freebsd.org/Graphics/OpenCL

The third change is the removal of Mesa 9.1.7 which was installed on
FreeBSD 9.3-RELEASE.  There is now only one version of Mesa in the Ports
tree (10.6.6) for all supported versions of FreeBSD.

Other, smaller changes:

* Include libosmesa into the Mesa framework; this changes libOSMesa
  shlib version.

* bsd.mesalib.mk was renamed and split up in two files namely
  Makefile.common and Makefile.targets.  So ports can overwrite variables
  set by Makefile.common and are used by Makefile.targets.

* Some text in the pkg-descr files was wrong, clean it up.  While here,
  update the WWW to the main mesa3d.org upstream page.

* devel/clinfo was added, a glxinfo like program but for OpenCL.

Non-x86 hardware reports are very welcome since we changed the framework
quite a bit.

Obtained from:	Graphics team development repo.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@397198 35697150-7ecd-e111-bb59-0022644237b5