sysutils/ipfs-go: IPFS implementation in Go
IPFS is a global, versioned, peer-to-peer filesystem. It combines good ideas
from Git, BitTorrent, Kademlia, SFS, and the Web. It is like a single
bittorrent swarm, exchanging git objects. IPFS provides an interface as simple
as the HTTP web, but with permanence built in.

WWW: https://github.com/ipfs/go-ipfs


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@397699 35697150-7ecd-e111-bb59-0022644237b5