Update to svn revsion r1704217.
Fix a regression in directory permissions and add a missing directory used
for shared extensions.  This combination of issues could cause incomplete
cleanup on pkg removal if OpenOffice had been run by a user in the wheel
group.

Add a note about the location of the Wiki Publisher extension to
pkg-message.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@397760 35697150-7ecd-e111-bb59-0022644237b5