graphics/djview4: port improvements
There are some improvements for graphics/djview4 port to make it usable on
desktop. For example, while there is a need to use register-djview-menu shell
script (on post-install stage, which is not available currently) to setup the
djview4 menu entries and icons, it is possible to install them manually, as
described in ${WRKSRC}/desktopfiles/README file. Also this application is
translated to different languages.

- add LICENSE_FILE
- add libtiff.so to LIB_DEPENDS as required shared library
- add desktop-file-utils to USES and sort it
- add linguist to USE_QT4, because there are available messages to translate
- add --disable-desktopfiles option to CONFIGURE_ARGS, which do not install
  scripts to register icons and menus
- replace CFLAGS and CXXFLAGS with CPPFLAGS for include directory
- improve sed patch by using & for replacement part
- strip web-plugin library, install icons and desktop files on post-install
  stage
- remove unneeded files after --disable-desktopfiles option from pkg-plist
- add translated files, icons and desktop files to pkg-plist

PR:		196409
Submitted by:	lightside <lightside@gmx.com>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@398050 35697150-7ecd-e111-bb59-0022644237b5