audio/deadbeef: fix GUI option handling
The audio/deadbeef port has a GUI radio option that allows to choose between
GTK2 and GTK3 sub-options. Apparently, choosing sub-options does not directly
affect the parent option, for example, choosing GTK2 doesn't turn the GUI
option. As it's used in PLIST_SUB, it could result in a broken package, e.g.
without gui even if GTK2 is enabled.

Fix that by adding appropriate _IMPLIES for sub-options and manually turn
off the parent option if none of sub-options was enabled.

Bump PORTREVISION so users could get a new package with GTK2 UI as it supposed
to be.

PR:		203974
Approved by:	vg (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@400331 35697150-7ecd-e111-bb59-0022644237b5