www/py-django-mptt: Update to 0.7.4, Fix *_DEPENDS
This change is made to fix the fact that the current version
of this port (0.6.0) is not compatible with Django 1.8, the
port that it RUN_DEPENDS on. It is broken at run-time in both
Ports HEAD and Quarterly.

A soon to be committed update to www/py-django-dpaste (D4048) requires
this update to also fix it being broken in the quarterly branch.

The only other consumer of this port www/py-django-cms, already
incorrectly depending on versions of Django and django-mptt that are
later than those specified in its setup.py:

  django-mptt>=0.5.1,<0.5.3',
  'Django>=1.4,<1.6',

- Update PORTVERSION and distinfo checksum (0.7.4)
- Update COMMENT to match upstream (setup.py:description=)
- Add LICENSE_FILE
- Django is a RUN not BUILD depends, remove accordingly
- Sort USE_PYTHON values (portlint)
- Enable achitecture independent (NO_ARCH)

Changes:

  https://github.com/django-mptt/django-mptt/compare/0.6.0...0.7.0

[1] https://svnweb.freebsd.org/changeset/ports/400452

Reviewed by:		antoine
Approved by:		portmgr (antoine)
MFH:			2015Q4 (pre-approved: antoine)
Differential Revision:	https://reviews.freebsd.org/D4024


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@400572 35697150-7ecd-e111-bb59-0022644237b5