Contract graphics/cal3d-devel into graphics/cal3d
The development of cal3d stopped in July 2006, about a week or two
after the last release (0.11.0).  It makes no sense to have both cal3d
and cal3d-devel, so let's copy over the -devel port over to the stable
one and remove the -devel port.  There's no reason to set a deprecation
period in this case.

While here:
  * remove conflicts
  * Update the WWW to the new home page

I've given graphics/cal3d a "fake" version of 0.11.1 to convey that it
is an update to the previous version of 0.11.0 rather than to completely
change the version scheme.

PR:		203135
Approved by:	maintainer timeout
Discussed:	portmgr (bapt)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@400654 35697150-7ecd-e111-bb59-0022644237b5